rename memorylimiter -> memorylimiterprocessor
