+++ processor/memorylimiterprocessor/memorylimiter.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package memorylimiter
+package memorylimiterprocessor
 
 import (
 	"context"
