+++ service/defaultcomponents/default_processors_test.go
@@ -28,7 +28,7 @@ import (
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/consumer/consumertest"
-	"go.opentelemetry.io/collector/processor/memorylimiter"
+	"go.opentelemetry.io/collector/processor/memorylimiterprocessor"
 )
 
 func TestDefaultProcessors(t *testing.T) {
@@ -47,7 +47,7 @@ func TestDefaultProcessors(t *testing.T) {
 		{
 			processor: "memory_limiter",
 			getConfigFn: func() config.Processor {
-				cfg := procFactories["memory_limiter"].CreateDefaultConfig().(*memorylimiter.Config)
+				cfg := procFactories["memory_limiter"].CreateDefaultConfig().(*memorylimiterprocessor.Config)
 				cfg.CheckInterval = 100 * time.Millisecond
 				cfg.MemoryLimitMiB = 1024 * 1024
 				return cfg
